refactor(observe): deduplicate MCPServerFilter into ObserveFilterBar
